moved vm_orchestrator from core
moved vm_orchestrator from core blocks
