+++ lib/vcloud.rb
@@ -14,6 +14,7 @@ require 'vcloud/core'
 require 'vcloud/config_loader'
 require 'vcloud/launch'
 require 'vcloud/query'
+require 'vcloud/vm_orchestrator'
 require 'vcloud/vapp_orchestrator'
 
 module Vcloud
