+++ lib/vcloud/core.rb
@@ -7,4 +7,3 @@ require 'vcloud/core/vdc'
 require 'vcloud/core/vm'
 require 'vcloud/core/vapp'
 require 'vcloud/core/vapp_template'
-require 'vcloud/core/vm_orchestrator'
