+++ lib/vcloud/vapp_orchestrator.rb
@@ -12,7 +12,7 @@ module Vcloud
 
           network_names = extract_vm_networks(vapp_config)
           vapp = Vcloud::Core::Vapp.instantiate(name, network_names, template_id, vdc_name)
-          Vcloud::Core::VmOrchestrator.new(vapp.fog_vms.first, vapp).customize(vapp_config[:vm]) if vapp_config[:vm]
+          Vcloud::VmOrchestrator.new(vapp.fog_vms.first, vapp).customize(vapp_config[:vm]) if vapp_config[:vm]
           vapp
         end
